+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 Changes:
 * Show displayname on chat instead of real name
+* Get message sender badge and color